some fixes

+++ src/calculator/Calculator.java
@@ -7,13 +7,13 @@ import interfaces.ISquare;
 public class Calculator implements ICalculator {
 
 	@Override
-	public float calcArea(IRectangle r) {
+	public double calcArea(IRectangle r) {
 		return r.getHeight() * r.getWidth();
 	}
 
 	@Override
-	public float calcArea(ISquare s) {
-		return (float) Math.pow(s.getSide(), 2);
+	public double calcArea(ISquare s) {
+		return Math.pow(s.getSide(), 2);
 	}
 
 }
